Merge complete

+++ Afya/app/src/main/java/com/example/afya/data/model/Trip.kt
@@ -6,7 +6,7 @@ import java.util.*
 
 class Trip(var num_sortie: Int?, var num_util: Int?, var date_sortie: String?,
            var heure_depart: String?, var heure_arrivee: String?, var lieu_depart: String?,
-           var distance_parcourue: Float?, var etapesByNum_sortie: List<Step>?
+           var distance_parcourue: Double?, var etapesByNum_sortie: List<Step>?
            ) : Serializable {
     fun formatDate() {
         val dateSortie = SimpleDateFormat("dd MMMM yyyy", Locale("fr")).format(
